Move __getitem__ to DhcpServiceInfo
